validation setup #1

+++ src/AYoutuber/Modules/LavFiltersInstallerExecuter.cs
@@ -26,25 +26,34 @@ namespace AYoutuber.Modules
         {
             base.Install(stateSaver);
 
-            string assemblyPath = this.Context.Parameters["assemblypath"];
+            this.LavFilterInstall();
+        }
 
-            var filterDirectory = Path.Combine(Path.GetDirectoryName(assemblyPath), "LavFilters");
-            var filterInstallerPath = Path.Combine(filterDirectory, "LAVFilters-Installer.exe");
-            ProcessStartInfo info = new ProcessStartInfo();
-            info.Arguments = " /VERYSILENT /SUPPERSSMSGBOXES";    // SILENT, VERYSILENT
-            info.FileName = filterInstallerPath;
-            info.WorkingDirectory = filterDirectory;
+        private int limitCount = 10 * 60 * 3;   // 3분동안 처리가 되는지 대기한다.
+        private int limitIndex = 0;
 
-            Process process = Process.Start(info);
-            int limitCount = 10 * 60 * 3;   // 3분동안 처리가 되는지 대기한다.
-            int limitIndex = 0;
-            while (!process.HasExited && process.Responding)
+        private void LavFilterInstall()
+        {
+            if (!File.Exists(filterInstallerPath))
             {
-                Thread.Sleep(100);
-                if(limitCount < limitIndex++)
+                string assemblyPath = this.Context.Parameters["assemblypath"];
+
+                var filterDirectory = Path.Combine(Path.GetDirectoryName(assemblyPath), "LavFilters");
+                var filterInstallerPath = Path.Combine(filterDirectory, "LAVFilters-Installer.exe");
+                ProcessStartInfo info = new ProcessStartInfo();
+                info.Arguments = " /VERYSILENT /SUPPERSSMSGBOXES";    // SILENT, VERYSILENT
+                info.FileName = filterInstallerPath;
+                info.WorkingDirectory = filterDirectory;
+
+                Process process = Process.Start(info);
+                while (!process.HasExited && process.Responding)
                 {
-                    process.Kill();
-                    throw new InvalidOperationException("LavFilter installation exception!");
+                    Thread.Sleep(100);
+                    if (limitCount < limitIndex++)
+                    {
+                        process.Kill();
+                        throw new InvalidOperationException("LavFilter installation exception!");
+                    }
                 }
             }
         }
@@ -53,41 +62,63 @@ namespace AYoutuber.Modules
         {
             base.Commit(savedState);
 
-            // 설치된 폴더 권한 조정
+            this.SetPermission();
+        }
+
+        private void SetPermission()
+        {
+            // 설치된 폴더의 Works 권한 조정
+            //string path = Path.Combine(Path.GetDirectoryName(this.Context.Parameters["assemblypath"]), "Works");
             string path = Path.GetDirectoryName(this.Context.Parameters["assemblypath"]);
+            if (!Directory.Exists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+
             DirectorySecurity sec = Directory.GetAccessControl(path);
             // Using this instead of the "Everyone" string means we work on non-English systems.
             SecurityIdentifier everyone = new SecurityIdentifier(WellKnownSidType.WorldSid, null);
             sec.AddAccessRule(new FileSystemAccessRule(everyone, FileSystemRights.Modify | FileSystemRights.Synchronize, InheritanceFlags.ContainerInherit | InheritanceFlags.ObjectInherit, PropagationFlags.None, AccessControlType.Allow));
+            SecurityIdentifier builtinUsers = new SecurityIdentifier(WellKnownSidType.BuiltinUsersSid, null);
+            sec.AddAccessRule(new FileSystemAccessRule(builtinUsers, FileSystemRights.Modify | FileSystemRights.Synchronize, InheritanceFlags.ContainerInherit | InheritanceFlags.ObjectInherit, PropagationFlags.None, AccessControlType.Allow));
             Directory.SetAccessControl(path, sec);
         }
 
         public override void Rollback(IDictionary savedState)
         {
+            this.LavFilterUninstall();
+
             base.Rollback(savedState);
         }
 
         public override void Uninstall(IDictionary savedState)
         {
+            this.LavFilterUninstall();
+
             base.Uninstall(savedState);
+        }
 
-            var filterDirectory = @"C:\Program Files (x86)\LAV Filters";
-            var filterInstallerPath = @"C:\Program Files (x86)\LAV Filters\unins000.exe";
-            ProcessStartInfo info = new ProcessStartInfo();
-            info.Arguments = " /VERYSILENT /SUPPERSSMSGBOXES";    // SILENT, VERYSILENT
-            info.FileName = filterInstallerPath;
-            info.WorkingDirectory = filterDirectory;
+        private string filterDirectory = @"C:\Program Files (x86)\LAV Filters";
+        private string filterInstallerPath = @"C:\Program Files (x86)\LAV Filters\unins000.exe";
 
-            Process process = Process.Start(info);
-            int limitCount = 10 * 60 * 3;   // 3분동안 처리가 되는지 대기한다.
-            int limitIndex = 0;
-            while (!process.HasExited && process.Responding)
+        private void LavFilterUninstall()
+        {
+            if (File.Exists(filterInstallerPath))
             {
-                Thread.Sleep(100);
-                if (limitCount < limitIndex++)
+                ProcessStartInfo info = new ProcessStartInfo();
+                info.Arguments = " /VERYSILENT /SUPPERSSMSGBOXES";    // SILENT, VERYSILENT
+                info.FileName = filterInstallerPath;
+                info.WorkingDirectory = filterDirectory;
+
+                Process process = Process.Start(info);
+                while (!process.HasExited && process.Responding)
                 {
-                    process.Kill();
-                    throw new InvalidOperationException("LavFilter uninstallation exception!");
+                    Thread.Sleep(100);
+                    if (limitCount < limitIndex++)
+                    {
+                        process.Kill();
+                        throw new InvalidOperationException("LavFilter uninstallation exception!");
+                    }
                 }
             }
         }
